+++ docs/assets/chevron-up-CGiOgYQB.js
@@ -1,0 +1,1 @@
+import{ht as e}from"./index-Dlbfwc3_.js";var t=e(`chevron-up`,[[`path`,{d:`m18 15-6-6-6 6`,key:`153udz`}]]);export{t};