+++ docs/assets/chevron-up-im90knuY.js
@@ -1,1 +1,0 @@
-import{ht as e}from"./index-LXaHvjDc.js";var t=e(`chevron-up`,[[`path`,{d:`m18 15-6-6-6 6`,key:`153udz`}]]);export{t};